requirements.txt
# Python dependencies for bootstrap scripts
PyYAML>=6.0
jsonschema>=4.0.0

